refactor(tree): move tree status utils
