+++ experimental/dds/tree2/src/core/tree/index.ts
@@ -39,6 +39,8 @@ export {
 	FieldUpPath,
 	compareUpPaths,
 	compareFieldUpPaths,
+	TreeStatus,
+	treeStatusFromPath,
 	UpPathDefault,
 } from "./pathTree";
 export {
