+++ experimental/dds/tree2/src/feature-libraries/editable-tree/editableField.ts
@@ -20,6 +20,8 @@ import {
 	ITreeCursor,
 	keyAsDetachedField,
 	rootField,
+	TreeStatus,
+	treeStatusFromPath,
 } from "../../core";
 import { FieldKind, Multiplicity } from "../modular-schema";
 import {
@@ -45,13 +47,11 @@ import {
 	isPrimitive,
 	keyIsValidIndex,
 	getOwnArrayKeys,
-	treeStatusFromPath,
 } from "./utilities";
 import { ProxyContext } from "./editableTreeContext";
 import {
 	EditableField,
 	EditableTree,
-	TreeStatus,
 	UnwrappedEditableField,
 	UnwrappedEditableTree,
 	proxyTargetSymbol,
