+++ experimental/dds/tree2/src/feature-libraries/editable-tree/utilities.ts
@@ -5,20 +5,13 @@
 
 import { assert } from "@fluidframework/core-utils";
 import { isStableId } from "@fluidframework/container-runtime";
-import {
-	DetachedField,
-	FieldKey,
-	TreeStoredSchema,
-	UpPath,
-	keyAsDetachedField,
-	rootField,
-} from "../../core";
+import { FieldKey, TreeStoredSchema } from "../../core";
 import { brand } from "../../util";
 import { valueSymbol } from "../contextuallyTyped";
 import { FieldKinds } from "../default-field-kinds";
 import { StableNodeKey } from "../node-key";
 import { getField } from "../untypedTree";
-import { EditableTree, TreeStatus } from "./editableTreeTypes";
+import { EditableTree } from "./editableTreeTypes";
 
 /**
  * @returns true iff `schema` trees should default to being viewed as just their value when possible.
@@ -105,34 +98,4 @@ export function getStableNodeKey(
 		);
 		return brand(id);
 	}
-}
-
-/**
- * Checks whether or not a given path is parented under the root field.
- * @param path - the path you want to check.
- * @returns the {@link DetachedField} which contains the path.
- */
-export function getDetachedFieldContainingPath(path: UpPath): DetachedField {
-	let currentPath = path;
-	while (currentPath !== undefined) {
-		if (currentPath.parent === undefined) {
-			return keyAsDetachedField(currentPath.parentField);
-		} else {
-			currentPath = currentPath.parent;
-		}
-	}
-	return keyAsDetachedField(path.parentField);
-}
-
-/**
- * Checks the path and returns the TreeStatus based on whether or not the detached field is the root field.
- * TODO: Performance: This is a slow initial implementation which traverses the entire path up to the root of the tree.
- * This should eventually be optimized.
- * @param path - the path you want to check
- * @returns the {@link TreeStatus} from the path provided.
- */
-export function treeStatusFromPath(path: UpPath): TreeStatus {
-	return getDetachedFieldContainingPath(path) === rootField
-		? TreeStatus.InDocument
-		: TreeStatus.Removed;
 }
