+++ experimental/dds/tree2/src/feature-libraries/index.ts
@@ -51,7 +51,6 @@ export {
 	comparePipeline,
 	compileSyntaxTree,
 	setField,
-	TreeStatus,
 } from "./editable-tree";
 
 export {
