+++ experimental/dds/tree2/src/index.ts
@@ -77,6 +77,7 @@ export {
 	MapTree,
 	LocalCommitSource,
 	forbiddenFieldKindIdentifier,
+	TreeStatus,
 } from "./core";
 
 export {
@@ -263,7 +264,6 @@ export {
 	compileSyntaxTree,
 	nodeKeyFieldKey,
 	SchemaLintConfiguration,
-	TreeStatus,
 	treeStatus,
 } from "./feature-libraries";
